docs/requirements.txt
fastapi
uvicorn
python-dotenv
starlette
python-multipart
requests

langchain>=0.3,<0.4
langchain-openai>=0.2,<0.3
langchain-community>=0.3,<0.4
langgraph>=0.2,<0.3
langchain-google-genai>=2.0,<3.0
langchain-text-splitters>=0.3,<0.4
langsmith>=0.1.140
chromadb>=0.5,<0.6

transformers
torch

pandas
beautifulsoup4
pypdf
cloudscraper
wikirate4py
name_matching
langdetect
PyPDF2
unstructured

paddlepaddle==2.6.2
paddleocr==2.9.1
opencv-python==4.10.0.84
pytesseract
Pillow==10.4.0
numpy==1.26.4

protobuf==3.20.2
google-generativeai
